ref(allocation_policies): extract match-block resolver into its own module
